+++ gatsby-ssr.js
@@ -36,6 +36,8 @@ function setColorsByTheme() {
     colorMode = prefersDarkScheme ? 'dark' : 'light';
   }
 
+  const invertedColorMode = colorMode === 'dark' ? 'light' : 'dark';
+
   const root = document.documentElement;
 
   root.style.setProperty(colorModeCSSProp, colorMode);
@@ -44,6 +46,7 @@ function setColorsByTheme() {
     const cssVarName = `--color-${name}`;
 
     root.style.setProperty(cssVarName, colorByTheme[colorMode]);
+    root.style.setProperty(`${cssVarName}-inverted`, colorByTheme[invertedColorMode]);
   });
 }
 
@@ -77,7 +80,8 @@ function FallbackStyles() {
   */
 
   const cssVariableString = Object.entries(COLORS).reduce(
-    (acc, [name, colorByTheme]) => `${acc}\n--color-${name}: ${colorByTheme.light};`,
+    (acc, [name, colorByTheme]) =>
+      `${acc}\n--color-${name}: ${colorByTheme.light};\n--color-${name}-inverted: ${colorByTheme.dark};`,
     '',
   );