+++ src/components/ThemeContextProvider.tsx
@@ -38,10 +38,13 @@ function ThemeContextProvider({ children }) {
       // TODO Uncomment the line if a dark mode toggle is added
       // localStorage.setItem(COLOR_MODE_KEY, newValue);
 
+      const invertedColorMode = newValue === 'dark' ? 'light' : 'dark';
+
       Object.entries(COLORS).forEach(([name, colorByTheme]) => {
         const cssVarName = `--color-${name}`;
 
         root.style.setProperty(cssVarName, colorByTheme[newValue]);
+        root.style.setProperty(`${cssVarName}-inverted`, colorByTheme[invertedColorMode]);
       });
 
       rawSetColorMode(newValue);